+++ .fvm/fvm_config.json
@@ -1,4 +1,3 @@
 {
-  "flutterSdkVersion": "3.13.9",
-  "flavors": {}
+  "flutterSdkVersion": "3.13.9"
 }
