+++ .fvmrc
@@ -1,0 +1,4 @@
+{
+  "flutter": "3.13.9",
+  "flavors": {}
+}
