+++ .gitignore
@@ -54,7 +54,6 @@ app.*.map.json
 *.env.default
 
 # Fvm
-.fvm/flutter_sdk
 
 # fastlane specific
 **/fastlane/report.xml
@@ -63,3 +62,6 @@ app.*.map.json
 **/fastlane/test_output
 .bundle/
 vendor/bundle/
+
+# FVM Version Cache
+.fvm/
