+++ integration_test/common/repository_mocks.dart
@@ -1,0 +1,4 @@
+import 'package:flutter_template/core/repository/session_repository.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockSessionRepository extends Mock implements SessionRepository {}
